RelWihDebInfo creates a debug version
Revert RelWihDebInfo to release. It is supposed to make the Debug symbols available, but there is not only that.
The MuseScore version it creates has the "Debug" menu and all debugging features. It is not suitable for normal users.
Took me a while to understand that it was not specific to my config.